Safari-609.1.11.1/LayoutTests/crypto/subtle/rsa-import-jwk-key-export-spki-key-expected.txt

Test importing a JWK RSA-OAEP public key with SHA-1 and then export it in SPKI format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
Exporting a key...
PASS bytesToHexString(spkiKey) is expectedSpkiKey
PASS successfullyParsed is true

TEST COMPLETE
